Added acc_bench_smm. Adjusted and improved Makefile. Fixed some compiler warnings.

+++ src/acc/opencl/smm/acc_opencl_smm.c
@@ -129,7 +129,7 @@ int libsmm_acc_transpose(const int* dev_trs_stack, int offset, int stack_size,
               new_config.wgsize = 0;
             }
           }
-          if (max_wgsize < new_config.wgsize || 0 == local) new_config.wgsize = 1;
+          if (max_wgsize < (int)new_config.wgsize || 0 == local) new_config.wgsize = 1;
           config = (config_t*)libxsmm_xregister(&key, sizeof(key), sizeof(new_config), &new_config);
         }
       }
@@ -139,7 +139,7 @@ int libsmm_acc_transpose(const int* dev_trs_stack, int offset, int stack_size,
     }
     assert((NULL != config && NULL != config->kernel) || EXIT_SUCCESS != result);
     if (EXIT_SUCCESS == result) {
-      const size_t work_size = (0 < config->wgsize ? config->wgsize : m) * stack_size;
+      const size_t work_size = (0 < config->wgsize ? config->wgsize : (size_t)m) * stack_size;
       ACC_OPENCL_CHECK(clSetKernelArg(config->kernel, 0, sizeof(cl_mem), ACC_OPENCL_MEM(dev_trs_stack)),
         "set batch-list argument of transpose kernel", result);
       ACC_OPENCL_CHECK(clSetKernelArg(config->kernel, 1, sizeof(int), &offset),
